backend/requirements.txt
# Django и основные зависимости
Django==5.0.1
djangorestframework==3.14.0

# WebSocket поддержка
channels==4.0.0
channels-redis==4.1.0

# База данных
psycopg2-binary==2.9.9

# OpenAI API
openai==1.6.1

# Валидация данных
pydantic==2.5.3

# Redis
redis==5.0.1

# CORS для разработки
django-cors-headers==4.3.1

# Переменные окружения
python-decouple==3.8

# Утилиты
python-dateutil==2.8.2

# Production сервер (опционально)
gunicorn==21.2.0
daphne==4.0.0